Produce ReusableBoxFuture and ReusableLocalBoxFuture with a macro
Adding new features in both is a bit troublesome because of code
duplication, so now we produce the duplicated code with a macro.
